Refactor d3 related functions to hook
